moved headlistener
